moved smurf methods to frontend-services
