+++ src/main/java/dk/kb/netarchivesuite/solrwayback/service/SolrWaybackResource.java
@@ -115,20 +115,6 @@ public class SolrWaybackResource {
   
  
   
-  @GET
-  @Path("smurf/tags")
-  @Produces({ MediaType.APPLICATION_JSON})
-  public  SmurfYearBuckets smurfNetarchiveTags( @QueryParam("tag") String tag , @QueryParam("fq") String filterQuery,  @QueryParam("startyear") Integer startyear) throws SolrWaybackServiceException {
-      try {                                                                                      
-        
-        if (startyear == null){
-          startyear=1990;
-        }
-        return Facade.generateNetarchiveSmurfData(tag, filterQuery,startyear);                  
-      } catch (Exception e) {         
-          throw handleServiceExceptions(e);
-      }
-  }
   
   
   @GET
@@ -143,20 +129,6 @@ public class SolrWaybackResource {
   }
  
   
-  
-  @GET
-  @Path("smurf/text")
-  @Produces({ MediaType.APPLICATION_JSON})
-  public  SmurfYearBuckets smurfNetarchiveText( @QueryParam("q") String q , @QueryParam("fq") String filterQuery,  @QueryParam("startyear") Integer startyear) throws SolrWaybackServiceException {
-      try {                                                                                                
-        if (startyear == null){
-          startyear=1990;
-        }
-        return Facade.generateNetarchiveTextSmurfData(q, filterQuery,startyear);                  
-      } catch (Exception e) {         
-          throw handleServiceExceptions(e);
-      }
-  }
   
 //Already removed to frontend. Keep until sure it is not used.
   @GET
